requirements.txt
Django~=6.0.2
python-decouple==3.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg[binary]
gunicorn
django-silk==5.4.3


